Moves dleq to new top-level zero-knowledge package.
