+++ oprf/client.go
@@ -4,7 +4,7 @@ import (
 	"crypto/rand"
 
 	"github.com/cloudflare/circl/group"
-	"github.com/cloudflare/circl/group/dleq"
+	"github.com/cloudflare/circl/zk/dleq"
 )
 
 type client struct{ params }
