+++ oprf/server.go
@@ -5,7 +5,7 @@ import (
 	"crypto/subtle"
 
 	"github.com/cloudflare/circl/group"
-	"github.com/cloudflare/circl/group/dleq"
+	"github.com/cloudflare/circl/zk/dleq"
 )
 
 type server struct {
