+++ oprf/vectors_test.go
@@ -13,8 +13,8 @@ import (
 	"testing"
 
 	"github.com/cloudflare/circl/group"
-	"github.com/cloudflare/circl/group/dleq"
 	"github.com/cloudflare/circl/internal/test"
+	"github.com/cloudflare/circl/zk/dleq"
 )
 
 type vector struct {
